Struct yup_hyper_mock::HostToReplyConnector

pub struct HostToReplyConnector {
    pub m: HashMap<String, String>,
}

A NetworkConnector which provides a single reply stream per host.

The mapping is done from full host url (e.g. http://host.name.com) to the singular reply the host is supposed to make.

Fields

m