Struct yup_hyper_mock::TeeConnector

pub struct TeeConnector<C> where C: NetworkConnector {
    pub connector: C,
}

A NetworkConnector embedding another NetworkConnector instance, and sets it up to write all reads and writes to standard error as well.

NOTE It was originally intended to allow arbitrary streams to copy data to, but I couldn't get passt the compiler with that as normal streams, like files, are not normally clonable. Maybe an Arc+Mutex would have helped ... .

Fields

connector